requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
spotipy==2.23.0
uritemplate==4.1.1
urllib3==2.2.1
